+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisRequest.cs
@@ -7,5 +7,19 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
     /// </summary>
     public class RunDynamicAnalysisRequest : OperationRequestBase
     {
+        /// <summary>
+        /// Unit: s (second).
+        /// </summary>
+        public double TimeStep { get; set; }
+
+        /// <summary>
+        /// Unit: s (second).
+        /// </summary>
+        public double FinalTime { get; set; }
+
+        /// <summary>
+        /// An additional information to be set in the file name that contains the analysis results.
+        /// </summary>
+        public string AdditionalFileNameInformation { get; set; }
     }
 }