+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisResponse.cs
@@ -5,7 +5,5 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
     /// <summary>
     /// It represents the response content of RunStaticAnalysis operation.
     /// </summary>
-    public class RunDynamicAnalysisResponse : OperationResponseBase<RunDynamicAnalysisResponseData>
-    {
-    }
+    public class RunDynamicAnalysisResponse : OperationResponseBase<RunDynamicAnalysisResponseData> { }
 }