+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisResponseData.cs
@@ -7,5 +7,9 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
     /// </summary>
     public class RunDynamicAnalysisResponseData : OperationResponseData
     {
+        /// <summary>
+        /// The full name of solution file.
+        /// </summary>
+        public string FullFileName { get; set; }
     }
 }