+++ Suspension/CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.2.0] - 2022-01-13
 ### Changed
  - Framework to NET6.0.
- - Nuget package MudRunner.Commons.Core to 1.1.0.
- - Nuget package MudRunner.Commons.DataContracts to 1.1.0.
+ - Nuget package MudRunner.Commons.Core to 1.2.0.
+ - Nuget package MudRunner.Commons.DataContracts to 1.2.0.
  - Renamed suspension A-Arm to wishbone in all application.  
  - Renamed PivotPoint1 to FrontPivot in all application.
  - Renamed PivotPoint2 to RearPivot in all application.
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Codes that was in MudRunner.Commons.
 ### Added
  - Operation CalculateSteeringKnuckleReactions. 
+ - Operation RunDynamicAnalysis.
+ - Class BasePaths with the base paths for application.
 
 ## [1.1.0] - 2021-12-20
 ### Added